+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
-## [2.0.0-beta1] - In progress
+## [2.0.0-beta2] - In progress
+
+### Added
+
+- New `YamlAdapter` adapter
+
+## [2.0.0-beta1] - 2021-04-07
 
 ### Added